Move isRecord to utils and remove deprecated opt
